Format about me section properly
About me section is the only one that is centered properly and the divs inside of that are also set to space between with a width.

Portfolio, Skills and Contact need to follow this same format.

If the width is less than the width set to About me section, it will just be centered. Width should not exceed the same amount about me section is set to.